+++ src/extension/listners.ts
@@ -2,10 +2,15 @@ import type { listnerType } from '@/extension/types'
 
 export const triggerListner = ( type: listnerType, listnerData: any ) => {
     const data = { type: "CLWALLET_EXT_LISTNER", data: { listner: type, data: listnerData } }
-    chrome.tabs.query({}, (tabs) => tabs.forEach( tab => 
-        {
+    chrome.tabs.query({}, (tabs) => tabs.forEach( async tab => 
+        {   
             if (tab?.id) {
-                chrome.tabs.sendMessage(tab.id, data) 
+                try {
+                    await chrome.tabs.sendMessage(tab.id, data)
+                } catch
+                {
+                    // ignore
+                }
             }
         }
     ));